Fixes improper outfit declarations (#2137)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
## About The Pull Request
A handful of ships did not declare some jobs correctly. This fixes all
instances of jobs declared simply as "jobtitle": # by replacing them
with a full outfit and slots declaration.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Job outfits not being declared in ship jsons causes the game to try and
grab an associated outfit on its own. Which can lead to improper outfit
loading.
This can be seen with tm's of #2034
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:phoaly
fix: Fixes improperly declared outfits in ship jsons
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->