Resprites Assorted Furniture! (#2140)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This PR redoes:

- The Rack

- The Holopad originally seen in /tg/station made by yours truly

- The Dresser


![image](https://github.com/shiptest-ss13/Shiptest/assets/77556824/8d4155b0-fa4d-43e5-8a67-dee001cb1d63)

All authored by yours truly!
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
These sprites were rather crusty looking, and the rack was a request to
update by @Bjarl, and I'm happy to help a mapper in need!
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: PositiveEntropy
imageadd: The Rack, Dresser and Holopad have been redone and updated!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->