The Most Serene Solar Confederation's Most Serene Solar Starship, The Chronicle (#2074)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
![2023-06-15 14 00
08](https://github.com/shiptest-ss13/Shiptest/assets/94164348/629b864a-e00f-481e-b9b9-03567d13d5e8)
![2023-06-15 14 00
10](https://github.com/shiptest-ss13/Shiptest/assets/94164348/825bb8b3-ab00-4643-9166-407a44145717)

It's The Solgov Ship. It uses The Solgov Content. 

More specifically it's a engineering / intelligence ship, kind of like a
mildly larger kansatsu. A lot of content relating to sensors still
doesn't exist, but they've got the gear and equipment to get the
authentic Solgov Experience. This includes, but is not limited to:
Energy halbreds, excessive amounts of blue and gold, paperwork, and
holopads.

Crew consists of
1 Captain
1 Overseer
2 Field Engineers
1 Sonnensoldner
2 Ship Engineers
1 Scribe
For a total of 8 crew.


## Why It's Good For The Game
Solgov content
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: table mounted intercoms now exist
add: The Chronicle Class, a Real Solgov Ship, is now in game
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->