Hyena nerf (#2091)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

![hyena](https://github.com/shiptest-ss13/Shiptest/assets/60533805/febc5a44-f3cd-42f9-8db3-c30bdc9b0c18)

Updates the Hyena. Changes are as follows:

1. Replaced syndicate cleavers with blastwave jackhammers
2. Downgraded scarlet hardsuits on miners to armored syndicate space
suits
3. Added access restrictions to engineering and foreman's office
4. Removed spare headsets from captain and foreman lockers for
consistency with other ships
5. Removed codespeak book
6. Pre-built R&D machines
7. Provided miners and engineer with jetpacks
8. Added functional turret control panel

## Why It's Good For The Game

The Hyena is yet another ship in need of touching up as mapping
standards shift, and its perpetual power scaling problems continue to
present themselves. The cleavers are a common complaint, so hopefully
making its mining gear somewhat more in line with other mining ships
will help. The Hyena still retains its other gimmicks, such as the bag
of explosives.

## Changelog

:cl:
tweak: Updated the Hyena with turret controls and visual updates
balance: Rebalanced the Hyena's mining gear
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->