Resprites Most Chairs! (#2160)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Hello everyone!

This PR updates most of the chairs to use new sprites made by me in
/tg/station originally, but with a twist! Most of them have been
re-shaded by a wiser yours truly, as well as keeping some of the
aesthetic ideas of the previous chairs!


![image](https://github.com/shiptest-ss13/Shiptest/assets/77556824/5999b68f-aeae-416c-a3b9-123aad7a3a6b)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The previous ones were wildly inconsistent, so hopefully this makes the
chairs more cohesive as well as providing a more clean aesthetic to our
server!
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: PositiveEntropy
imageadd: Redoes a good majority of chairs!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->