Atmos sprite update (#2136)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Updates the sprites for the atmos firesuit, inhand fireaxe, and fireaxe
cabinet to fit in with the rest of the engineering sprites. Also,
fireaxe item sprite is no longer upside down when unwielded.

![image](https://github.com/shiptest-ss13/Shiptest/assets/93578146/24f75111-dd9c-42f5-b35d-e52e53ad03fa)

![image](https://github.com/shiptest-ss13/Shiptest/assets/93578146/93b1e744-f0b9-4119-9142-17e5a664ea87)

Sprites are from tgstation/tgstation#72736 and tgstation/tgstation#65199

## Why It's Good For The Game
Makes atmos firesuit and axe sprites fit in with the newer engineering
sprites, also I think it looks nice.

## Changelog

:cl:
imageadd: updates some atmos sprites to fit in with the rest of
engineering
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->