Organizes lists in surgery files, nerfs ghetto surgery a bit (#2080)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
I'm not a big fan of people just whipping out a toolbox or whatever, and
being able to fully revive people. What's the point of medical tools
and/or ships if it's effectively fully available, to everyone?
* Removes pen from tend wounds
* Rebalances ghetto surgery success/fail
* Brain / Heart will now decay in around 45 minutes.
* Every other organ will decay in around 30 minutes.

The gist of it is this, in order of effectiveness:
* Scalpel analogue is Kitchen Knife / Sharp Things
* Retractor is Screwdriver
* Hemostat is Wirecutters
* Saw is Fireaxe, Hatchet, Sharpthings
* Cautery is Welder / Hot Things

Other than tool changes, expect the following:
* 40% surgery success chance with the above tools is baseline for most
operations.
* Failing a step will do brute damage, break a bone, or other side
effects I can come up with
* The chances can still be increased with ethanol, sterilizine, gloves,
beds, etc. (I think this is how it works)

## Why It's Good For The Game
A greater importance should be placed on having medical equipment versus
just a toolbelt. I would prefer to see all of it go but that would
require a lot more work on the medical system.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
del: You can no longer tend wounds with a pen
balance: Ghetto surgery is less reliable across the board.
balance: Organs will now decay in ~30 minutes, heart and brain in ~45
minutes.
balance: Inducers now work for revival surgery.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->