Make it so you can't use any item to cancel mechanical surgery (#1989)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Right now if you click an ipc surgery with anything that is not the
right item it "screw" the shell and force you to restart teh surgery
which is a pain. (it applies to EVERY item like even gause).
You can still close the shell with a knife if you don't have a
screwdriver.

## Why It's Good For The Game

Said above

## Changelog

:cl:
tweak: Mechanical surgery won't cancel at any wrong move.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: ritorizo <41831966+ritorizo@users.noreply.github.com>
Co-authored-by: ritorizo <ritorizo@localhost>
Co-authored-by: Mark Suckerberg <mark@suckerberg.gay>