Removes the Metis (#2192)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Removes the Metis from the game.

## Why It's Good For The Game

The Metis has been slated for removal for months now, pending a
replacement. A replacement has not been forthcoming, and, in the
meanwhile, the Metis continues to cause problems. If nothing else,
removing it should hopefully encourage a prompt replacement.

## Changelog

:cl:
del: Removed the Metis
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->