Solarian DLC: SSG-669C and SolGov Captain's Cloak! (#2178)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Title!

This PR adds the missing SSG-669C Sniper Rifle to the game! Enjoy a 5
round bolt action weapon with all that armor piercing caseless goodness!


![image](https://github.com/shiptest-ss13/Shiptest/assets/77556824/dce69d7d-1ec4-420a-ae72-ecf3a93e4d23)

This PR also adds a new cosmetic item for Solarian Captains, a cloak!


![image](https://github.com/shiptest-ss13/Shiptest/assets/77556824/cacc38ed-3679-48e7-8a0f-f395b9274a9b)

Now you can successfully style on non-solarians!
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
One was a missing item that was meant to be added in the original
solarian armory PR, and the other is a fun little cosmetic addition for
players to enjoy!
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: PositiveEntropy, Retlaw34
add: Adds the solarian SSG-669C Sniper Rifle!
add: Adds a cloak for the SolGov captain!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Imaginos16 <77556824+Imaginos16@users.noreply.github.com>
Co-authored-by: thgvr <81882910+thgvr@users.noreply.github.com>