Sanitizes syndicate headsets in outfits (#2189)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Removes Syndie headsets from Syndie job outfits that are not officers or
Twinkleshine crew, bringing the Syndie headset distribution in line with
other factions. Does not touch the maps themselves, so if spare headsets
need to be removed (and they should) this will have to be done
separately.

## Why It's Good For The Game

Faction radio channels are intended to function as something of a
faction-exclusive wideband, not a common channel for every member of the
faction to flood with minutiae. This is the same reason ships with
pre-built comms do not relay the common channel across the sector- Comms
spam sucks.

## Changelog

:cl:
balance: Removed Syndie headsets from non-officer jobs for parity with
other factions
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->