Maintains the Riggs + mini flavor tweaks (#2181)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Brings the Riggs up to par, primarily by adding a holofield to the cargo
bay. Additionally, some minor aesthetic and QoL updates were made.

- Adds holofield to cargo bay
- Replaces shuttle displays with ship viewscreens
- Tweaks wallmount positions for aesthetics
- Replaces the 10mm surplus rifle with a Winchester

## Why It's Good For The Game

keeping old ships up to snuff is always useful

## Changelog

:cl:
tweak: Updated the Riggs with a holofield and some visual polish
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->