Outposts v0.2: Test PR (#1789)
## About The Pull Request

ships spawn in (non-exclusive, so don't get too comfortable) hangars at
the outpost sized to them, connected to a new "main level" of the
outpost via an elevator. new hangars are generated as necessary as ships
spawn / dock. let me know if radio works across hangars or not

~~this is mega-unfinished and not fit for merge, but it IS fit for test,
and that's what we do around here isn't it???~~

i'm basically done with this. that took a hot minute.
for detailed documentation, check here:
https://hackmd.io/@Xegx-gsTR2WyYII99ZKtrg/B1updZdK2

## Why It's Good For The Game

it's cool

## Changelog

:cl:
add: Ships now spawn at the outpost, in hangars. Take the elevator in
the top-right down to the main level.
add: 100% guaranteed most realistic ss13 elevator experience if you
ignore all the things that aren't realistic
/:cl:

---------

Co-authored-by: Bjarl <94164348+Bjarl@users.noreply.github.com>
Co-authored-by: spockye <79304582+spockye@users.noreply.github.com>