Updates Headset Message Options (#2198)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request


![imagen](https://github.com/shiptest-ss13/Shiptest/assets/75212565/9752fe8c-ebc7-4aa4-8178-aa1230ee864f)


![imagen](https://github.com/shiptest-ss13/Shiptest/assets/75212565/0de32684-82ef-4f39-865a-321ca5d7f747)


![imagen](https://github.com/shiptest-ss13/Shiptest/assets/75212565/5f4a33cb-af56-4032-b10f-299360bdb554)

Adds more options to the headset message verb so admins can message
players' respective faction without relying on CentCom or narrating,
such as Inteq Risk Management, the Syndicate's many branches, and the
Minutemen.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
We have more factions than just the Syndicate and NanoTrasen now, and
it's best to use something more specific than CentCom for contacting
players on, say, a CMM ship.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
admin: There are more options to headset messages, including Inteq, CMM,
and multiple Syndicate branches.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: meemofcourse <75212565+meemofcourse@users.noreply.github.com>
Co-authored-by: Mark Suckerberg <mark@suckerberg.gay>