Minor Gun and Moth Fixes (#2167)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Brings the Toy C-20r smg in line with all of the foam weapons by giving
it the same firing sound as the rest. Removes the ability for it to be
suppressed for consistency as well. Changes "colorfull" in the rainbow
moth's description to the proper spelling of colorful.

## Why It's Good For The Game

Load up the Halftrack. Test out my foam guns and not have my eardrums
hurt from an overly loud "Toy" C-20r. Hug my favorite moth plushie buddy
and not cry at the extra l in their description. Life's good.

## Changelog

:cl:
tweak: Removed the ability for the Toy C-20r to accept a suppressor.
fix: Gave the Toy C-20r the proper firing sound.
spellcheck: Fixed the description of the rainbow moth.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Geoengi <139566513+Geoengi@users.noreply.github.com>