Paracelsus-class Definitive Edition (#2210)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Wow! Back so soon?

Yeah so turns out, I messed up. I admit it. The Paracelsus-class was
shipped too soon and I realized I made plenty of mistakes. Thankfully
however, rectifying them was very achievable!

This PR changes the layout of the Paracelsus by quite a significant
margin, adding an extended cargo bay with a small cargo office on it, as
well as changing the blast doors to be double cycling holofields!

![2023-07-30 19 42
21](https://github.com/shiptest-ss13/Shiptest/assets/77556824/71642138-559c-43c0-9b01-7accc3752c0f)

![2023-07-30 19 42
27](https://github.com/shiptest-ss13/Shiptest/assets/77556824/684928dc-f7bc-4b25-9e36-04654d3df39a)

One fun addition to this ship has been a psychologist room for some
quality mental evaluations of patients!


![image](https://github.com/shiptest-ss13/Shiptest/assets/77556824/a9ff3dd5-49c7-486b-a03c-384469a2cfc7)

Speaking of psychologists and patients, new roles!

- Bureaucrat has been replaced by an Overseer, with an appropriate
locker for them in the bridge.
- A new Field Engineer has been added to the fray!
- A dedicated Solarian Psychologist now exists on the Paracelsus-class!
- Bespoke solarian Patient roles now exist!

And finally, three new clothing options for solarian ships have been
added!
- The Solarian Suit
- The Solarian Hat
- The Formal Solarian Suitskirt
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
I hope these changes are better now :)
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: PositiveEntropy
add: Adds new SolGov clothes for people to enjoy!
add: Bespoke Psychologist and Patient roles for SolGov exist now!
fix: The Paracelsus is now a proper ship that isn't a good 50% empty
space anymore!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->