Twink Tweaks (#2209)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Removes the SecTech, fixes a couple of extinguishers, adds a variant of
the Syndicate mask that has a VC, so you don't see syndicate crews
walking around with horse masks.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Vouches can be redeemed in sectech to get NT weapons. Also, you don't
see syndicate crews walking around with horse masks.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Syndicate Voice Changer masks.
tweak: Removed the SecTech from the Twinkleshine, and changed the masks
that the crews get.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->