re:Saint-Roumain Militia  (#2247)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Redoes the SRM clothing

## Why It's Good For The Game


![Screenshot_6736](https://github.com/shiptest-ss13/Shiptest/assets/58402542/2f1adb23-4bbb-4b9f-b4ac-33d8d6d318e0)

## Changelog

:cl:
add: resprites SRM gear
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->