Access restricts the Aegis bridge (#2238)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Does what it says on the tin- adds access restrictions and finer door
controls to the Aegis's bridge. The Aegis still has a larger rework on
the horizon when SUNS content is finished, but for now this is just a
quick fix to a recurring issue.

## Why It's Good For The Game

Bridges, as a rule, should be access restricted to officer roles on any
ship of sufficient size so it doesn't just turn into another assistant
den. The Aegis is no exception.

## Changelog

:cl:
tweak: Added access restrictions to the bridge on the Aegis
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->