Slightly more robust holopads (#2135)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Makes it so holocalls aren't interrupted by the target shuttle docking,
or the hologram leaving the projection range. It'll just send them to
the source tile if either happen. Also tries to slow down movement a
little, but there's only so much I can do.

## Why It's Good For The Game

Holopads are probably one of the most important communication methods in
the game, since it's one of the few reliable inter-ship ones.

## Changelog

:cl:
tweak: Holocalls now persist even if the target ship docks or the
hologram leaves the projection area.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->